Add husky pre-push check
Add husky dependency & configuration.
Since we automatically cut a new release when one is detected in a commit, we need to run the tests before pushing, to ensure that our rules are still applied as expected.
Add notes to README as to why we do not have any other integrations.