Add TS lint testing and @typescript-eslint/explicit-module-boundary-types warning (#24)
We want to use @typescript-eslint/explicit-module-boundary-types and will have it as a warning for now to not disrupt us as much. Later on we will make it an error.
This is the first TS rule we are adding in this configuration, therefore, the ts eslint config and a ts file for testing local eslint issues are now part of testing.