Update md file and add package-lock.json
In the last pr, the master build failed because eslint-config-frontier-react updated after the pr was made, but before it was merged, and that update included a change to the test.

Add package-lock.json to see changes better and not have a pr's checks be possibly different than the master's checks due to deps changing things between a pr creation and merge.

@fs/eslint-config-frontier-react@11.4.2 turned off you-dont-need-lodash-underscore/clone-deep and that made the test fail on master.