Support esx from eslint-config-frontier-react (#32)
esx is applied to src files since that is what end users use and that is
what we need to restrict. See
https://icseng.atlassian.net/browse/FRONTIER-2569

esx is a way to restrict from using specific new es features so that
users on older browsers do not get errors as much.

See
https://github.com/fs-webdev/eslint-config-frontier-react/pull/62/files
for the eslint-config-frontier-react pr.